refactor(main): extract bootstrap and logging
